moved log and data dir setup to utils
